Fix ovs/netlink interface discovery for internal ovs interface
When receiving an ovs update try to lookup interfaces in the netns
scopde with the mac address first, if not found try to lookup interfaces
of type openvswitch in the netns scope since the interface could have
been added by the netlink driver. If so reset the uuid to the ovsdb
uuid.