Agents are now able to update central graph
The graph events are converted to GraphMessage which are sent to
the analyzer through a websocket connection. The analyzer unmarshal the
GraphMessage(s) and re-apply them to the local graph. A Resync mechanism is
there to resend all the local graph of an agent to the analyzer when an
connection error occurs. The same Re-Sync mechanism is used for the
WebUI.