Improves flow capture reporting and fix reconnection issue
Attach flow capture by sFlow to the bridge where the sFlow agent is
registered and not anymore to one of the interface in the bridge.
Shows the interface with the status set to DOWN in red color.

This patch also fixes several issues when there was a disconnection
between either the analyzer and agents or between analyzer and the
WebUI.