Introduce common websocket server/client
The common websocket server is used for
the graph and the alert but also for the
future flow server.

Now uses only one Websocket endpoint removing
ping/pong for each endpoint. A namespace attribute
in WSMessage is used to make the difference between
different messages endpoint.

Clients announce themself to the server with an Hello
message so that the server can register them with
their hostname and is able to send them messages
using the host to point them.

Change-Id: Iccf6f301320f7c16aef20a55565731be8bda4a02