neutron: Update lookup criteria for Neutron ports.
When we have a VM attached to an OVS interface, we need to
use the "attached-mac" external_id instead of the interface MAC
address.  The "attached-mac" is the address of the interface as seen
in the VM and is what matches the MAC address on the Neutron port.

Now that we have the OVS interface external_ids available, the
"iface-id" external_id will include the Neutron port UUID, so we can
use that as the most effective search criteria.

Change-Id: Ic5a7ae5f78c1495caa9b8d31dd35d21fd2499a7a
Signed-off-by: Russell Bryant <rbryant@redhat.com>