graph: introduce a gremlin traversal mechanism
This patch introduces a subset of the gremlin language
aiming to provide a generic way to do requests on graph.

This patch also improves the way of the metadata elements
are compared.

The functional tests will be modified to leverage this
new way of retrieving nodes/edges in a follow-up
patch.

Change-Id: Id30e0d296d13ba1d20793f22746f46b65d227fe4