[flows] Fixup UUID and document flow UUID,ID usage
flow.UUID is unique in the universe, as it should be used as a key of an
hashtable. By design 2 different flows, their UUID are always different.
flow.UUID can be used as Database Index.

flow.TrackingID could be used to identify an unique flow whatever it has
been captured on the infrastructure. flow.TrackingID is calculated from
the bytes of the first packet of his session.
flow.TrackingID can be used as a Tag.

Change-Id: I1c6638934e19c28a6af4350df4ac9edff60a22f8