flow: do not expose tickers
There is no need to expose the internal
mechanic outside the flow table.
This patch add three functions to the flow
table in order to start the flow processing
,to stop it and to flow flush it.

This patch fixes the pcap probe which was not
using the async way of processing the flows

A follow-up patch will come in order to add
an allocator of flow tables. This will help
to provide a flow API.

Change-Id: I17d2a66c6178eb8dbf71a04789c33a0c5e075abf