flow: introduce table allocator and a query mechanism
The allocator will be used in order to
interact will the flow tables.

This patch remove all the Flush method that
were used in probes and use the allocator
for that purpose.

This patch also add a table query mechanism
allowing to do request on flow tables
directly on through the allocator. In the
case of the allocator it will aggregate the
replies. A first use of this feature if search
flows based on ProbeNodeUUID.

Change-Id: If69d7426267619d040983c487d5754b54ead156a