flow: use node uuid instead of graphpath
This patch use uuid to reference a node in
the flows instead of graphpath. Gremlin now
allows us to find a node by its uuid.
Moreover since we are going to introduce the
history of the graph a same graphpath could
represent two different nodes. Using uuid
we are sure to associate correctly the right
nodes to the flows.

Change-Id: I94c1a6b66ec273e4ae60567ef0e1d8f6e9baceaa