flow: fixes and refactoring helping to introduce more probes
This patch fixes issues when having truncated packet
which is the case when capturing using sFlow.

Fixe metrics reported.

This patch now handles IPV6 and add key to the UUID
hash computation to be sure that 2 different keys don't
lead to the same UUID due to 2 way of hashing headers.

This patch simplify the way the flow are handled and
expose some functions that will be used for incoming
probes.

Change-Id: If3cf697fc03de516bf6792d5331328b209886b6c