flow: fix Sort/Dedup/Limit order and optimize Dedup
this patch fixes an issue with

Sort().Dedup() and Dedup().Sort() and optimize
Dedup doing depuplication during merge.

It also fixes Limit when G.Flows().Limit()
and add non reduced Sort step.

Rename Sorted in protobuf message as it is a
request not a state of the data.

Change-Id: Ide6ca68eb888b922388a1774b390b1a021102cb4
Reviewed-on: http://softwarefactory-project.io/r/5092
Reviewed-by: Sylvain Baubeau <sbaubeau@redhat.com>
Tested-by: Sylvain Baubeau <sbaubeau@redhat.com>
Workflow: Sylvain Baubeau <sbaubeau@redhat.com>
Tested-by: Jenkins CI <jenkins@softwarefactory-project.io>