Manages captures from the analyzer
This patch moves part of the capture handling from the agents
to the analyzer. The latter is now responsible for listening
etcd. When a new capture is created/deleted or a new node
matching a capture appears, the analyzer sends a message to the
through WebSocket.

Change-Id: Ia39a7c053605f7afa81a1f2a3c03cd2013de7cab
Reviewed-on: https://softwarefactory-project.io/r/5428
Reviewed-by: Sylvain Afchain <safchain@gmail.com>
Tested-by: Sylvain Afchain <safchain@gmail.com>
Workflow: Sylvain Afchain <safchain@gmail.com>
Tested-by: Jenkins CI <jenkins@softwarefactory-project.io>