netlink: only create sockets inside the namespace
Before this patch, the netlink probe would do all its work inside
a network namespace. This can cause the Golang runtime to create
threads for other goroutine that belongs to the namespace. To
avoid this issue, we switch to the namespace only to create sockets
for both netlink and ethtool and immediately leave the namespace.

Change-Id: I9dd88cc03da8287a4da091a116d9bcf5d637afa4
Reviewed-on: https://softwarefactory-project.io/r/5472
Reviewed-by: Sylvain Afchain <safchain@gmail.com>
Tested-by: Sylvain Afchain <safchain@gmail.com>
Workflow: Sylvain Afchain <safchain@gmail.com>