capture: refactor to LockOSThread the afpacket goroutine
It seems to be better to lock the afpacket goroutine
on the thread that opened the namespace.

This patch embed the afpacket google packet with
a poll timeout fix in order to avoid a memory
corruption when closing the handle. We will remove it
once the proposed PR will be merged.

Change-Id: I29848b2984b7fa2739c9117e633ac3f4ea63dac7
Reviewed-on: https://softwarefactory-project.io/r/5629
Reviewed-by: Sylvain Baubeau <sbaubeau@redhat.com>
Tested-by: Sylvain Baubeau <sbaubeau@redhat.com>
Workflow: Sylvain Baubeau <sbaubeau@redhat.com>
Reviewed-by: Nicolas PLANEL <nplanel@redhat.com>
Tested-by: Nicolas PLANEL <nplanel@redhat.com>
Workflow: Nicolas PLANEL <nplanel@redhat.com>