alert: switch condition to JavaScript query
The condition can be a pure Gremlin or a block of
JavaScript call. The method 'Gremlin' is available
in JavaScript to execute other Gremlin queries.

The condition result is evaluated to true if it returns:
  - true
  - a non empty string
  - a number different from zero
  - a non empty array
  - a non empty map

This commit also allows the user to specify an action
to execute when the alert is triggered, that can be either:
- a webhook. A POST request is issued with the JSON message as payload
- a local file (prefixed by file://) to execute a script. It receives the JSON
  message through stdin

The 'Trigger' attribute of an alert can now be specified
whether the alarm should be evaluated when an event occurred
on the graph or periodically.

Change-Id: Ia7e16b60d84818bada2b10e9890d0e678d518a0c
Reviewed-on: https://softwarefactory-project.io/r/5887
Reviewed-by: Sylvain Afchain <safchain@gmail.com>
Tested-by: Sylvain Afchain <safchain@gmail.com>
Workflow: Sylvain Afchain <safchain@gmail.com>
Tested-by: Jenkins CI <jenkins@softwarefactory-project.io>