flow: use a chan to push packets to FlowTable
This patch gets ride of the "Default" behavior of
the flow table. The flow are not created anymore by
probes but by the flow table itself. Probes generate
packets and push them in a dedicated chan of the
flow table.

This allows to simplify all the packet processing and
to remove timeout mechanismes at the probe levels.

Move all the flow related function at a unique place,
functions which was in statistics but was basically
parsing layers.

Change-Id: Ia746b10e4175826b2fd9b7f4bb0f9b37d9501f11
Reviewed-on: https://softwarefactory-project.io/r/5700
Reviewed-by: Sylvain Baubeau <sbaubeau@redhat.com>
Tested-by: Sylvain Baubeau <sbaubeau@redhat.com>
Workflow: Sylvain Baubeau <sbaubeau@redhat.com>