test: convert SFlowWithPCAP test into Unit test
Before this patch the test was creating a topology
only to start a "fake" sFlow capture to be able to
use the replay tool to send packets on the opened sFlow
port.

This patch simplies the test, just reading the pcap
trace.

It also fixes the metrics checked that were wrong.

Change-Id: I5abb27ca7cf2550ca7bd1ccba8af9fd038560542
Reviewed-on: https://softwarefactory-project.io/r/5883
Reviewed-by: Sylvain Baubeau <sbaubeau@redhat.com>
Tested-by: Sylvain Baubeau <sbaubeau@redhat.com>
Workflow: Sylvain Baubeau <sbaubeau@redhat.com>
Tested-by: Jenkins CI <jenkins@softwarefactory-project.io>