test: add Contrail encapsulation capture
We add a new decoder to decode the payload of a MPLS layer. This
decoder tryies to decode the payload as IPv4, IPv6 and fallback to
ethernet. Note this decoder is needed since gopacket only decodes IPv4
and IPv6 in MPLS packets.

The trace contains two packets sniffed by one capture, on one
interface. One packet layerpath is Ethernet/IPv4/UDP/MPLS/IPv4/ICMPv4
while the other one is Ethernet/IPv4/UDP/MPLS/Ethernet/IPv4/TCP.

Contrail can remove the ethernet header of the packet generated by a VM
when it pushes it into the tunnel. In particular, if source and
destination IPs don't belong to the same network, Contrail removes the
ethernet header. So, the MPLS payload can be ethernet or IP.

Change-Id: Ic625ed5a5cc506cc79f1b30920b73931e4696416
Reviewed-on: https://softwarefactory-project.io/r/5940
Reviewed-by: Sylvain Afchain <safchain@gmail.com>
Tested-by: Sylvain Afchain <safchain@gmail.com>
Workflow: Sylvain Afchain <safchain@gmail.com>
Tested-by: Jenkins CI <jenkins@softwarefactory-project.io>