topology: add peering probe which creates a link based on MAC
This new probe is loaded by analyzer and create a L2 link
according to MAC addresses and a specific metadata.

Neutron make use of this new probe in order to create a
link between VM tap. This patch removes the use of fabric
fabric for that purpose.

Change-Id: I48218692db389b19f35f7f7e5ab8bb1198a8800a
Reviewed-on: https://softwarefactory-project.io/r/6002
Reviewed-by: Sylvain Baubeau <sbaubeau@redhat.com>
Tested-by: Sylvain Baubeau <sbaubeau@redhat.com>
Workflow: Sylvain Baubeau <sbaubeau@redhat.com>