neutron: fix Neutron calls issue
A previous attempt to reduce the number of Neutron calls was too
efficient: we are no longer able to get Neutron metadatas in all cases!

For instance, we first get an update on the MAC attributes. The cache
entry was then set to PendingPortMetadata and the Neutron probe failed
to get metadatas. Later, the Contrail probe sets the attribute
Neutron/iface-mac which is used by Neutron. But, since
PendingPortMetadata has been set, the Neutron call is not realized.

In this commit, before a call to Neutron, we set the cache entry of port
metadatas (metadatas can be zeros). We only do another Neutron call if
port metadatas have changed.

Change-Id: Ib873eac286111c7a8e04c84fecc7c77335042ea3
Reviewed-on: https://softwarefactory-project.io/r/6066
Reviewed-by: Sylvain Afchain <safchain@gmail.com>
Tested-by: Sylvain Afchain <safchain@gmail.com>
Workflow: Sylvain Afchain <safchain@gmail.com>