ui: don't redraw topology on every node update
Since node statistics are updated regularly the topology
gets refreshed for every node update.

In the case of updated metadatas are not useful to the topology
display we should skip the redraw.

Also, if the selected node is updated we need to redraw it
with the correct class and size.

Change-Id: I95df89b139dfa96a802fcb1c74faf5d181f639a5
Reviewed-on: https://softwarefactory-project.io/r/6568
Tested-by: Jenkins CI <jenkins@softwarefactory-project.io>
Reviewed-by: Sylvain Afchain <safchain@gmail.com>
Tested-by: Sylvain Afchain <safchain@gmail.com>
Workflow: Sylvain Afchain <safchain@gmail.com>