test: update scale script
usage: scale.sh <num analyzers> <num agents> <num vms>

This new version allows to simulate
multi-agent/analyzer environments.

Each agent is started in a namespace
simulating a host.

All the host are connected throught a ovs bridge.

Each agent host hosts some VMs. All the VMs are
connected to a local ovs within the host with
a local instance of ovsdb.

The total number of VMs namespaces is equal to
<num agents> * <num vms>

All the agent ovs are connected in a full mesh
way to the others thanks to a VXLAN tunnel. So
that VMs can ping each other.

Change-Id: I5fe63b7fad4473ec697863903fea0c8dab478c5d
Reviewed-on: https://softwarefactory-project.io/r/6702
Reviewed-by: Sylvain Baubeau <sbaubeau@redhat.com>
Tested-by: Sylvain Baubeau <sbaubeau@redhat.com>
Workflow: Sylvain Baubeau <sbaubeau@redhat.com>