flow: remove analyzer flow table
This patch removes the flow table from analyzers.
There is no need to keep this flow table as it is
never requested.

Moreover it leads in having different results
in term of metrics since we have different
update timing between analyzers and agents.

Use a LRU cache for enhancement and move the
pipeline in the flow table.

This patch also removes temporaly the discovery
and the conversation tabs as they need to use gremlin
instead of specific api json encoder.

Change-Id: I37e6b37a30704f51e9c44f6fb18b7fb28c337f72
Reviewed-on: https://softwarefactory-project.io/r/6817
Reviewed-by: Sylvain Baubeau <sbaubeau@redhat.com>
Tested-by: Sylvain Baubeau <sbaubeau@redhat.com>
Workflow: Sylvain Baubeau <sbaubeau@redhat.com>