graph: add UpdatedAt attribute
Commit 874f69f409b82e893ac2ae438c3cacf62500080b changed the semantics
of CreatedAt and DeletedAt for nodes and edges. They are now exposed
to the user, who can expect to have the same Created for all the
revisions of a node/edge. This patch introduces a new 'UpdatedAt'
attribute, which is used for time queries in addition to Created and
DeletedAt.

This patch also convert all dates from seconds to milliseconds for
more precision

Change-Id: I6a120ca3bc96182cb7223878f2a845e042a559e7
Reviewed-on: https://softwarefactory-project.io/r/6855
Reviewed-by: Sylvain Afchain <safchain@gmail.com>
Tested-by: Sylvain Afchain <safchain@gmail.com>
Workflow: Sylvain Afchain <safchain@gmail.com>
Tested-by: Jenkins CI <jenkins@softwarefactory-project.io>