sort: introduce sort for nodes and use it for metrics test
This patch introduces sort for nodes and leverages this for
metrics fixes. We always need to sort just before getting
metrics because it is useless to do it before as only the
first nodes of a gremlin expression are sorted and
this maybe not the nodes used later for the metrics, ex:

V().Has('Name', 'vm1').out()

Only the V() nodes has been sorted not the final ones
which could have totally different time.

Change-Id: I4b6454b3c8ded1d81c919b2dad71ceb98aadbaac
Reviewed-on: https://softwarefactory-project.io/r/7049
Reviewed-by: Masco Kaliyamoorthy <mkaliyam@redhat.com>
Workflow: Sylvain Afchain <safchain@gmail.com>
Tested-by: Jenkins CI <jenkins@softwarefactory-project.io>