afpacket: back to CGO
Since the last gopacket update it doesn't
rely any more in CGO syscall which causes
issue for some reasons. Until we find the
reasons of these issues we use CGO version.

In order to keep the BPF support this patch
makes a mix of the CGO version and the BPF
support added meanwhile.

Change-Id: I3f363edbc965ab64c5ed462d3bb539f3de265fc9
Reviewed-on: https://softwarefactory-project.io/r/7094
Reviewed-by: Masco Kaliyamoorthy <mkaliyam@redhat.com>
Workflow: Masco Kaliyamoorthy <mkaliyam@redhat.com>
Tested-by: Jenkins CI <jenkins@softwarefactory-project.io>