graph: replace timestamp mechanism by archive mechanism
This patch aims to fix histo issue introducing a
Revision field and supporting multiple updates
at the same time.

For elasticsearch two updates will be merged.
For orientdb the Revsion field will be used
to ensure to get the right version of an element.

This patch ensures that the timestamp are the ones
set by agents.

This patch ensures also that edges deleted due to
node delete will have the same DeletedAt.

Change-Id: Iac56efe88d7d0c83997210a8ab2f727405fe5a70
Fixes: #250