flow: add raw packet to flow
This patch add a new field to flow
for raw packets.

Raw packets can be captured and stored
according to the rawpacket-limit parameter
of capture.

RawPackets() step can be used to retrieve
raw packets. The pcap output format can be
used to generate a pcap file.

Add a picto to the flow details in order to
download pcap file of raw packets.

Change-Id: I077f1484c6f9c9d90a3b169292c7acccc64bcf48