netlink: fix orphaned VLAN interfaces
Skydive does not show the relationship between Linux VLAN interfaces and
their parent interfaces. This is because the netlink topology probe
looks up the Vlan metadata field as a String, when adding the
child/parent relationship, but in fact Vlan is an integer.

This change fixes this issue by explicitly converting the Vlan metadata
field to an int64, and using the appropriate GetField_int64() function
to look up the field.

fixes: #304