neutron: do not reuse metadata map as value.
We shouldn't use a map as metadata value or
we shouldn't modify it after commit since
a map is a pointer modifying it can be reported
to the datastore especially in case of bulk
operations.

This patch increases a bit the error chan.

It also run the test against the datastore.