openflow: handles superimposed openflow rules
Rules can have the same filters (table, cookie, field filter)
but different priorities. They cannot be distinguished by ovs-ofctl monitor
and only dump-flows can discriminate them. The code was failing to
accomodate those rules by trying to be too smart both for adding and
deleting rules. The solution is to create a bucket for all rules with
same filter and associate the UUID obtained by forgetting about the priority.
Updating the content of the bucket is made by comparing the content
of the graph and the result of the query. Usually there is only one
rule in the bucket so performances are good.

closes #367