gremlin: add subgraph step
SubGraph returns a new graph based on previous steps.
E or V step can be used on this new Graph.

The goal is to provide a way to get a restricted view
of the Graph. For ex: only the "layer2" topology.

This allows to do:

G.E().Has('RelationType', 'layer2').SubGraph()

or

G.V().Has('Type', 'veth').SubGraph()

Related: #417