netlink: makes metric real object and get ride of TimedMetric
This aims to filter out empty metrics
in order to reduce the amount of data
exchanged.

This patch renames also statistics to
metric to keep thing in sync with flows.

There is no anymore TimedMetric all the
Metrics have to implement an only one
interface Metric. This aims to improve
the serialization performance.