elastic: Make sure the system doesnt get stuck due to overloaded ES DB
We add in the case of topology DB updates currently for ES only
channel that will make bulk updates to storage and clear the update buffer
and clear the buffer in case and we reach some watermark. Reaching it will mean
that ES doesnt stand the frequency of updates.

Conflicts:
	config/config.go