Use only trollius and not asyncio on both python versions
Using asyncio for python3 and trollius for python2 requires us to have
both packages in our dependency list (requirements.txt and setup.py).
When using automated installers, they may parse both packages as
required, resulting in unusable python2 environment (asyncio is not
compatible with python2).

As trollius is compatible with both versions, it would be easier to use
it on both platforms instead.