Added login method to python client
Added login method to the python client to authenticate with the skydive
analyzer.
The authentication should be performed before the main loop is started.
The method returns a boolean specifying if the login succeeded. In case
it did, a cookie is kept and is sent with all the future requests.

Future work: detect cases in which the authentication is expired, and
raise a notification to the user or re-authenticate automatically.

Co-Authored-By: Omer Anson <oaanson@gmail.com>