auth: get ride of http basic auth
This patch removes the support of the HTTP
part of our basic auth support as it causes
UI issue and authentification path inconsistency.

With this patch only the login endpoint using
the cookie returned as token is available.

Since this commit 81a691913310c2c019eb475bd64a760d8f6a6859
the python library relies on login and cookie.