elasticsearch: refactoring leveraging internal elasticsearch
This patch changes the way the records are handles:

For the topology:

Two indexes: one live and one archive
Only the archived index uses the rolling index
mechanism. The live index keeps track of current
objects. When a graph element is deleted. It is
deleted from the live index and insert as deleted
in the archive index. The previous version of a
graph element are kept in a cache and store in the
archive index on the next graph element update.

For the flows:

All the indexes are using the rolling index mechanism.