opencontrail: add routing table in Contrail metadatas
This commit adds the Contrail routing table in the Contrail metadata
of a node. It relies on the Contrail binary `rt` to be notified on
route updates.

The routing table is added in the Contrail metadata and not in already
existing routing table attributes because it seems hard to maintain
the routing table with two producers: the netlink probe squashes the
whole routing table metadata on each route updates.