websocket: marshal message before broadcasting and remove bulk
In order to avoid race we need to marshal message before pushing
them into a chan. As clients can use different protocol we can
marshal before pushing into it. Moreover there is not need to
push into a broadcast chan that will read messages to push them
into the client chan.

The bulk mechanism is also removed as not anymore used.