docker/ci: cross-compiler - use static
As we can't cross compile a dynamic skydive executable for now
we're limited to a static executable for cross-architecture.

ref: https://bugs.launchpad.net/ubuntu/+source/glib2.0/+bug/1784779
and https://bugs.launchpad.net/ubuntu/+source/util-linux/+bug/1784772

This static executable doesn't require the qemu trick to run a
non-native apt-get to build a dynamic docker environment so we've
reverted that change for now for now.