agent/host: Using lshw to report hardware
This commit is about using lshw tooling to report the hardware
information of nodes.

The JSON output of lshw is parsed and a little bit adjusted to rename
'children' nodes into user-readable types.

As lshw provides much more cpu info that the actual code,
the existing CPU reporting is only used when lshw is not available.

Signed-off-by: Erwan Velu <erwan@redhat.com>