netlink: fix link state reporting
When setting a switch port to down, the kernel reports the interface
as down. The netlink message has the flag IFF_UP set to up. Skydive
was using this field to determine the state of the interface. The
correct way is to use the operation state of the netlink message.

Reported-by: Erwan Velu <erwan@redhat.com>