libvirt: initial version of libvirt probe
Libvirt probe dynamically polls the status of libvirt domains and
their network interfaces. For each new interface, it tries to enrich
the coresponding tun device with the name of the domain, the guest mac
and the complete address of the device on the guest pci bus.

The probe also adds one node per domain that tracks the status of the
libvirt domain and the association to interfaces.

The probe relies on libvirt-go the official go binding for libvirt
that requires the c libvirt library. It is compiled only under Linux.

closes #1410
Co-authored-by: Pierre Crégut <pierre.cregut@orange.com>
Co-authored-by: Sergey Glazyrin <s.glazyrin@partner.samsung.com>