ansible: initial support for podman
Background:
libpod aka Podman is library for running OCI-based containers in Pods.
It is likely going to replace Docker in newer versions of RHEL systems,
which means Docker won't be supported in the future.

The CLI is supposed to work like it did with Docker (with some changes);
so this patch is pretty basic. It adds skydive_container_cli variable
(default to docker for backward compatibiility) and replace the
hardcoded "docker" commands by using this variable instead.

Some adjustments were made:
- podman doesn't need docker socket access.
- podman doesn't support --restart natively, we need a systemd service.
- podman doesn't require to have docker started.

Future iterations will be made to support proper SystemD/Podman
integration like it was done in TripleO for other containers:
http://my1.fr/blog/openstack-containerization-with-podman-part-2-operations/

Closes #1517
Signed-off-by: Emilien Macchi <emilien@redhat.com>