ebpf: introduce update flow operation
This patch aims to reduce cpu and memory usage.
It introduces a way to send flow update operation
from the eBPF probe to the flow table so that only
the metrics are sent over the chan. The flow is not
fully parsed anymore only metrics that are coming
from the kernel space are used.