sriov: initial support for sriov - netlink/pci part
Uses PCI address to associate an interface with the correct virtual function.
The interface is tagged with the ID of the VF.

Waits until num_vfs and Vfs structure in netlink converge. There is a race
in Linux kernel. Changing the value of num_vfs triggers an event before the
value is actually available in num_vfs for other processes.

test adapted to find the sriov interface and check it is up.