sriov: libvirt part
Creates nodes for PCI passthrough interfaces so that the sriov virtual function
can be associated to something. This node will also be used by the peering
probe if a skydive agent is run in the VM.

It should also work for physical interfaces handled to a VM or virtual interfaces
managed by vfio-pci.