bugfix: use ifindex instead of names for links key
Because of Linux consistent network device naming, interfaces are
renamed but links does not really handle this conversion
that well. Using ifindex solves this problem as it will not
change during the lifespan of an interface (and if an interface is
recreated, it will probably get a new name - tested with sriov).