graph: ensure edges are properly deleted when a node is deleted
Previously, we did not delete them because we were relying on the agent
to send OnEdgeDeleted messages. This was wrong as, in addition to the
graph being inconsistent before the OnEdgeDeleted messages are received,
the links may not exist on the agent. If the edges were created by an
analyzer probe, they would never be deleted resulting in edge with missing
parent and/or child