contrib/exporters: Add localfiles storer for debugging
Storing transformed flow objects in a local directory eases debugging
process for exporters.

To use this storer, use the following configuration for the exporter
main process:

    pipeline:
      store:
        type: localfiles
        localfiles:
          base_dir: /tmp/flows
          max_seconds_per_stream: 3600

The directory structure underneath the `base_dir` is identical to the
one used in S3 (Object Store).

Each file is accompanied by a `.metadata.json` file which holds the
fields that are stored as metadata headers in S3.