contrib/exporters/secadvisor: Encode flows in a JSON object
During the refactoring to pipelines the export format for flows was
changed to JSON array, whereas Security Advisor expects the objects in
COS/S3 to be a JSON object with a top-level "data" key whose value is an
array of flow objects.

This change adds an `sa` encode type which adds this top-level element
(and uses the core JSON encoder to perform the actual encoding, so
`pretty: true` still works).

To use this encoder, the secadvisor exporter should have this config:

    pipeline:
      encode:
        type: sa
        json:
          pretty: false   # or true