contrib/exporters/secadvisor: Fix IP's name caching
Name resolving for IP addresses depend on the IP address and the TID in
which the flow was captured.  When caching the result, we now use both
IP and TID in the cache key (an IP address might appear in different places
in the cluster and refer to different containers/hosts/...).