analyzer: Allow setting startup capture type
When using a startup capture, the config allowed setting the capture's
gremlin (which nodes should capture) and BPF filter string, but the
capture type was hard-coded to "pcap".

This change allows setting the type from the configuration: either
setting the `analyzer.startup.capture_type` in the YAML config file or
by setting the SKYDIVE_ANALYZER_STARTUP_CAPTURE_TYPE environment
variable.

Possible values for `capture_type` are those that appear in `ProbeTypes`
(see `common/capture.go`); they are listed in a comment in the example
config file; a value of "" (which is the default) will let Skydive
choose the capture type according to the node type.