Elasticsearch: define several hosts and allow invalid certs
Allow skydive to connect to Elasticsearch servers with TLS but invalid
certs.

For Elasticsearch clusters, be able to only define a single endpoint
creates a SPoF at start time. Allow to pass an array of endpoints.
This breaks the current config, which expect an string in the key
"host".
Now is an slice in the key "hosts".

Expose configuration for basic auth (previously supported with
http://user:password@...)

Add option 'disable_sniffing' to etc/skydive.yml.default, already
implemented.

Add option 'disable_healthcheck' to etc/skydive.yml.default.

Fix a typo in the error string when Elasticsearch version could not be
obtained.

Add olivere/elastic logging to skydive logger. Ease debuging problems
with Elasticsearch.