Fix node patch tests
Tests were failing because when a patch is ignored, server returns a 304
code without body.
Table tests have been modified to detect this case.
CurdClient.Update method now returns the return code of the server.
We also check that all the complete node returned by the call to patch
is exactly the same as if we get that node later.
Node revision were incorrectly set in the tests, it should be increased
if a patch is applied correctly.

Changed table tests name to filter these easily.