Elasticsearch: store prevRevision when fetching nodes/edges
When skydive fetchs nodes and edges at boot time from elasticsearch,
those nodes/edges were not being saved in the b.prevRevision map.

If later we tried to change the metadata of that nodes/edges (using
MetadataUpdated) it failed because it does not recognise that node/edge.

This commits modifies the GetNodes and GetEdges functions to store into
b.prevRevision nodes/edges fetched from Elasticsearch.